Implement gen* function generator
Add :frequency kwarg to fcompat and add to with-gen* arguments.

fcompat can generate functions that validate the arguments and fabricate a
return value compliant with the :ret of the :old and :new specs, ignoring the
:fn specs if present. This mirrors fspec's gen*.

Passing frequency to with-gen* fixes the case if gfn is nil or false it'll throw
a NullPointerException trying to access the frequency map.